if default images dont delete them

+++ controllers/userController.js
@@ -201,7 +201,9 @@ const changeDisplayPicture = async (req, res) => {
     const prevDpName = req.userInfo.displayPicture;
     const newDpName = req.file.filename;
 
-    await fs.unlink(`public/dp/${prevDpName}`);
+    if (prevDpName !== 'femaleDP.png' && prevDpName !== 'maleDP.png') {
+      await fs.unlink(`public/dp/${prevDpName}`);
+    }
 
     const updatedUser = await User.findByIdAndUpdate(
       { _id: req.userId },
